+++ classes/Generator/Generator/Builder.php
@@ -31,58 +31,16 @@ class Generator_Generator_Builder
 	protected $_generators = array();
 
 	/**
-	 * The pretend mode to be applied to each generator
-	 * @var boolean
+	 * The global values to be applied to each generator
+	 * @var array
 	 */
-	protected $_pretend;
-
-	/**
-	 * The force mode be applied to each generator
-	 * @var boolean
-	 */
-	protected $_force;
-
-	/**
-	 * The verify mode be applied to each generator
-	 * @var boolean
-	 */
-	protected $_verify;
+	protected $_globals = array();
 
 	/**
 	 * Is the builder ready to be executed?
 	 * @var boolean
 	 */
 	protected $_is_prepared = FALSE;
-
-	/**
-	 * Parameter defaults to be applied to each generator
-	 * @var array
-	 */
-	protected $_defaults = array();
-
-	/**
-	 * The absolute base path under which each item should be created
-	 * @var string
-	 */
-	protected $_path;
-
-	/**
-	 * The module folder under which each item should be created
-	 * @var string
-	 */
-	protected $_module;
-
-	/**
-	 * Overrides the view template file used by each generator
-	 * @var string
-	 */
-	protected $_template;
-
-	/**
-	 * Overrides the views template directory used by each generator
-	 * @var string
-	 */
-	protected $_template_dir;
 
 	/**
 	 * The main factory method for returning new builder instances.
@@ -187,11 +145,8 @@ class Generator_Generator_Builder
 	 */
 	public function with_defaults(array $defaults = NULL)
 	{
-		if ($defaults)
-		{
-			$this->_defaults = $defaults;
-			$this->_is_prepared = FALSE;
-		}
+		$this->_globals['defaults'] = $defaults;
+		$this->_is_prepared = FALSE;
 
 		return $this;
 	}
@@ -205,9 +160,9 @@ class Generator_Generator_Builder
 	 */
 	public function with_pretend($pretend = TRUE)
 	{
-		$this->_pretend = (bool) $pretend;
-
+		$this->_globals['pretend'] = (bool) $pretend;
 		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -218,7 +173,7 @@ class Generator_Generator_Builder
 	 */
 	public function is_pretend()
 	{
-		return $this->_pretend;
+		return Arr::get($this->_globals, 'pretend') === TRUE;
 	}
 
 	/**
@@ -230,9 +185,9 @@ class Generator_Generator_Builder
 	 */
 	public function with_force($force = TRUE)
 	{
-		$this->_force = (bool) $force;
-
+		$this->_globals['force'] = (bool) $force;
 		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -245,9 +200,9 @@ class Generator_Generator_Builder
 	 */
 	public function with_verify($verify = TRUE)
 	{
-		$this->_verify = (bool) $verify;
-
+		$this->_globals['verify'] = (bool) $verify;
 		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -260,9 +215,9 @@ class Generator_Generator_Builder
 	 */
 	public function with_path($path)
 	{
-		$this->_path = (string) $path;
-
+		$this->_globals['path'] = (string) $path;
 		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -276,9 +231,9 @@ class Generator_Generator_Builder
 	 */
 	public function with_module($module)
 	{
-		$this->_module = (string) $module;
-
+		$this->_globals['module'] = (string) $module;
 		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -291,9 +246,9 @@ class Generator_Generator_Builder
 	 */
 	public function with_template($template)
 	{
-		$this->_template = (string) $template;
-
+		$this->_globals['template'] = (string) $template;
 		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -306,9 +261,27 @@ class Generator_Generator_Builder
 	 */
 	public function with_template_dir($path)
 	{
-		$this->_template_dir = (string) $path;
-
+		$this->_globals['template_dir'] = (string) $path;
 		$this->_is_prepared = FALSE;
+
+		return $this;
+	}
+
+	/**
+	 * Lists the global values that are to be applied to each generator type added
+	 * by the builder using the with_* methods, or sets them via a passed array.
+	 *
+	 * @param   array  $globals  The list of global values to be set on generators
+	 * @return  array|Generator_Builder  List of stored globals or this instance
+	 */
+	public function globals(array $globals = NULL)
+	{
+		if ($globals === NULL)
+			return $this->_globals;
+
+		$this->_globals = $globals;
+		$this->_is_prepared = FALSE;
+
 		return $this;
 	}
 
@@ -329,7 +302,7 @@ class Generator_Generator_Builder
 	 * Allows inspection of the current generators list for debugging purposes.
 	 *
 	 * @param   boolean  $rendered  Should rendered output be displayed?
-	 * @return  array
+	 * @return  array  The generators list
 	 */
 	public function inspect($rendered = TRUE)
 	{
@@ -449,13 +422,13 @@ class Generator_Generator_Builder
 		foreach ($this->_generators as $generator)
 		{
 			// Set the module for the generator, if any
-			$generator->module($this->_module);
+			$generator->module(Arr::get($this->_globals, 'module'));
 
 			// Set the verify mode for the generator
-			$generator->verify($this->_verify);
+			$generator->verify(Arr::get($this->_globals, 'verify'));
 
 			// Set the custom base path for the generator, if any
-			$generator->path($this->_path);
+			$generator->path(Arr::get($this->_globals, 'path'));
 
 			if ( ! $generator->file())
 			{
@@ -463,14 +436,17 @@ class Generator_Generator_Builder
 				$generator->guess_filename();
 			}
 
-			// Builder defaults should be merged with the generator defaults
-			$generator->defaults(array_merge($generator->defaults(), $this->_defaults));
+			if (isset($this->_globals['defaults']))
+			{
+				// Builder defaults should be merged with the generator defaults
+				$generator->defaults(array_merge($generator->defaults(), $this->_globals['defaults']));
+			}
 
 			// Set the other global options
-			$generator->template_dir($this->_template_dir);
-			$generator->template($this->_template);
-			$generator->pretend($this->_pretend);
-			$generator->force($this->_force);
+			$generator->template_dir(Arr::get($this->_globals, 'template_dir'));
+			$generator->template(Arr::get($this->_globals, 'template'));
+			$generator->pretend(Arr::get($this->_globals, 'pretend'));
+			$generator->force(Arr::get($this->_globals, 'force'));
 		}
 
 		// We're finished preparing